% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countryStats}
\alias{countryStats}
\title{Calculates stats for a given country}
\usage{
countryStats(country, games)
}
\arguments{
\item{country}{A country string, such as "DEU"}

\item{games}{The data loaded by 'loadData'}
}
\description{
Calculates stats for a given country
}
\examples{
NA
}

